requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cleanup==5.2.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
filter==0.0.0.20200724
Markdown==3.3.4
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
